% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-cpp.R
\name{benchmark_dose_tmb}
\alias{benchmark_dose_tmb}
\title{Semi-parametric benchmark dosing}
\usage{
benchmark_dose_tmb(
  monosmooths,
  smooths,
  linearterms,
  data,
  exposure,
  response,
  x0,
  p0,
  BMR,
  verbose = FALSE,
  eps = 1e-06,
  maxitr = 10,
  bayes_boot = 1000
)
}
\arguments{
\item{monosmooths}{A list of \code{mgcv}-compatible smooth formula terms of the type \code{s(x,bs='bs')} to be treated as monotone smooths. Must
use B-Spline bases. Must contain the variable specified in \code{exposure}.}

\item{smooths}{A list of \code{mgcv}-compatible smooth formula terms of the type \code{s(x,bs='bs')} to be treated as non-monotone smooths. Must
use B-Spline bases.}

\item{linearterms}{A one-sided formula of the form \code{~x1+x2} compatible with \code{model.matrix}, defining any linear terms in the model.}

\item{data}{\code{data.frame} containing the variables in \code{smooths}, \code{monosmooths}, and \code{response}.}

\item{exposure}{Character vector naming the exposure variable in the \code{monosmooths} formulas. Will always be treated as monotone decreasing.}

\item{response}{Character vector naming the response variable in \code{data}.}

\item{x0}{Baseline exposure, default \code{0}}

\item{p0}{Baseline response}

\item{BMR}{Benchmark response}

\item{verbose}{Logical, print progress and diagnostic information for debugging? Default \code{FALSE}.}

\item{eps}{Numerical tolerance for Newton's method}

\item{maxitr}{Maximum number of iterations for Newton's method.}

\item{bayes_boot}{Nonegative integer, how many Bayesian "bootstrap" iterations to do; means how many posterior draws to base
sample-based inferences for the BMD(L) off of.
Must be a positive integer.}
}
\description{
Benchmark dosing based on a custom implementation of monotone B-spline-based additive models
using TMB and Rcpp
}
\note{
The BMDL assigned to the \code{bmdl} output object slot, and accessed via \code{get_bmd()} etc,
is based on Bayesian bootstrapping
}
