% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{bounded_newton}
\alias{bounded_newton}
\title{Reflected Newton's method}
\usage{
bounded_newton(g, bounds, eps = 1e-06)
}
\arguments{
\item{g}{Function to find the root of. Passed to \code{rlang::as_function}.}

\item{bounds}{Numeric vector of length 2 containing lower and upper bounds for where the root is}

\item{eps}{Numerical tolerance}
}
\value{
A numeric value representing the root.
}
\description{
R function implementing Newton's method for root finding with reflection
}
\details{
This function solves \code{g(x)=0} for a solution satisfying \code{x>=min(bounds)},
\code{x<=max(bounds)}, and \code{|g(x)|<eps}.
}
\examples{
bounded_newton(function(x) x^2,c(-1,1))
bounded_newton(~.x^2,c(-1,1))

}
