% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_computation_times}
\alias{get_computation_times}
\alias{get_relative_computation_times}
\title{Get computation times}
\usage{
get_computation_times(object, ...)

get_relative_computation_times(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Named numeric vector containing the computation times in seconds.
}
\description{
Helper function to extract the computation time in seconds for each step of what's
estimated in an object of class \code{semibmd}
returned by \code{benchmark_dose}.
}
