% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{estimate_bmd}
\alias{estimate_bmd}
\title{Semi-parametric benchmark dosing}
\usage{
estimate_bmd(
  formula,
  data,
  exposure,
  x0 = 0,
  p0 = 0.05,
  BMR = 0.05,
  monotone = TRUE
)
}
\arguments{
\item{formula}{formula passed to \code{scam::scam} or \code{mgcv::gam}}

\item{data}{\code{data.frame} passed to \code{scam::scam} or \code{mgcv::gam}}

\item{exposure}{Character vector naming the exposure variable in \code{formula}}

\item{x0}{Baseline exposure}

\item{p0}{Baseline response}

\item{BMR}{Benchmark response}

\item{monotone}{Logical: should a monotone generalized additive model be used as the dose-response model?
Default \code{TRUE} uses \code{scam::scam}; setting to \code{FALSE} uses \code{mgcv::gam}.}
}
\description{
Benchmark dosing based on semi-parametric (monotone) generalized additive models
}
