% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_estimates}
\alias{get_estimates}
\title{Get parameter estimates from the fitted dose-response model}
\usage{
get_estimates(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose_tmb}}

\item{...}{Not used}
}
\value{
A list containing elements \code{beta}, \code{gamma}, \code{alpha}, \code{lambda}, and \code{sigma}
representing, respectively: the un-restricted regression coefficients, the monotone transformed regression coefficients,
the intercept, the estimated smoothing parameters, and the estimated residual standard deviation, from the fitted dose-response
model. The parameter Hessian is not included, because we are very careful about the scale on which we apply asymptotic
Gaussian approximations in this problem. For standard errors, we recommend bootstrapping; see \code{plot.semibmd}.
}
\description{
Helper function to extract the parameter estimates from the fitted dose-response model in an object of class \code{semibmd}
returned by \code{benchmark_dose_tmb}.
}
