% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_all_bmdl}
\alias{get_all_bmdl}
\title{Get all benchmark dose lower limits}
\usage{
get_all_bmdl(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Named numeric vector containing the BMDLs. Currently this includes \code{score}
and \code{delta}.
}
\description{
Helper function to extract all BMDL's estimated in an object of class \code{semibmd}
returned by \code{benchmark_dose}.
}
