% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_approximations}
\alias{get_approximations}
\title{Get quantities related to the asymptotic approximations}
\usage{
get_approximations(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Named numeric vector containing the approximate quantities. Currently this includes \code{Vn} (variance of \code{U(xb)})
and \code{Upn} (derivative of \code{U} at estimated BMD).
}
\description{
Helper function to extract variance and estimated derivative used in the construction
of BMDLs.
}
