% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_bmd}
\alias{get_bmd}
\title{Get the benchmark dose and lower limit}
\usage{
get_bmd(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Numeric vector of length 2 containing \code{c(bmd,bmdl)}.
}
\description{
Helper function to extract the bmd and bmdl from an object of class \code{semibmd}
returned by \code{benchmark_dose}.
}
