% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_errors}
\alias{get_errors}
\title{Check for errors in BMD estimation}
\usage{
get_errors(object, verbose = FALSE, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{verbose}{Logical, print any errors obtained? Default \code{FALSE}.}

\item{...}{Not used}
}
\value{
Logical, were any errors obtained during the fitting process? If \code{TRUE} and \code{verbose=TRUE},
will print the errors
}
\description{
Check whether errors were returned by the estimation procedure for the model fitting,
BMD or BMDL calculation.
}
