% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_model}
\alias{get_model}
\title{Get the dose-response model}
\usage{
get_model(object)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}
}
\value{
Object of class \code{scam} or \code{gam} containing the fitted dose-response
model used to compute the bmd(l).
}
\description{
Helper function to extract the dose-response model object, of class \code{scam}
or \code{gam}, from an object of class \code{semibmd}
returned by \code{benchmark_dose}.
}
