% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_uxb}
\alias{get_uxb}
\title{Check the estimated BMD}
\usage{
get_uxb(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Value of the nonlinear equation \code{U(x)} evaluated at \code{x = } estimated benchmark dose.
If the BMD is estimated correctly, this should equal \code{0}.
}
\description{
Check the value of the nonlinear equation defining the estimated benchmark dose
at the estimated benchmark dose. Used for debugging and diagnosing BMD estimation
failure.
}
