% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_uxb}
\alias{get_uxb}
\alias{get_psixl}
\title{Check the estimated BMD/L}
\usage{
get_uxb(object, ...)

get_psixl(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd}
returned by \code{benchmark_dose}}

\item{...}{Not used}
}
\value{
Value of the nonlinear equation \code{U(x)} evaluated at \code{x = } estimated benchmark dose.
If the BMD is estimated correctly, this should equal \code{0}.
Similar for \code{Psi}.
}
\description{
Check the value of the nonlinear equation defining the estimated benchmark dose (lower)
at the estimated benchmark dose (lower). Used for debugging and diagnosing BMD/L estimation
failure.
}
