% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{plot.semibmd}
\alias{plot.semibmd}
\title{Plot method for \code{semibmd} objects}
\usage{
\method{plot}{semibmd}(x, ...)
}
\arguments{
\item{x}{Object of class \code{semibmd} returned by \code{semibmd::benchmark_dose}.}

\item{...}{Not used.}
}
\value{
Prints a plot with the dose-response and benchmark dose information on it.
}
\description{
Plot method for \code{semibmd} objects returned by \code{semibmd::benchmark_dose}.
Combines the standard plot from \code{scam/gam} with the benchmark dosing information
computed by \code{benchmark_dose} and \code{summary.semibmd}.
}
\details{
This function first calls \code{summary.semibmd}, then calls \code{plot.scam/gam}
as appropriate and modifies the resulting plot with the benchmark dosing information.
}
