% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{predict.semibmd}
\alias{predict.semibmd}
\alias{fitted.semibmd}
\alias{residuals.semibmd}
\title{Prediction methods for \code{semibmd} objects}
\usage{
\method{predict}{semibmd}(object, newdata = get_data(object), ...)

\method{fitted}{semibmd}(object, ...)

\method{residuals}{semibmd}(object, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd} returned by \code{semibmd::benchmark_dose}.}

\item{newdata}{Data frame containing all the variables in \code{get_data(object)} and one row for each value for
which a predicted response is required.}

\item{...}{Not used.}
}
\value{
A numeric vector containing the predictions or residuals as appropriate.
}
\description{
Predict, fitted, and residuals methods for \code{semibmd} objects returned by \code{semibmd::benchmark_dose}.
Combines the standard plot from \code{scam/gam} with the benchmark dosing information
computed by \code{benchmark_dose} and \code{summary.semibmd}.
}
\details{
The \code{predict} function computes the design matrix \code{X} at \code{newdata} and then computes \code{X \%*\% gamma} where
\code{gamma} is the estimated parameter vector for the monotone regression coefficients, denoted by \code{beta\_c} in the paper.
Then, the mean of these predictions is subtracted and the estimated intercept \code{alpha} is added on.
For more advanced computations required to predict each fitted smooth, see \code{plot.semibmd} with \code{plot = FALSE}.
The \code{fitted} function just calls \code{predict} with its default arguments and the \code{residuals} function just subtracts
\code{fitted} from the response.
}
