% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reflect}
\alias{reflect}
\title{Reflection for bounded newton's method}
\usage{
reflect(yt, bounds)
}
\arguments{
\item{yt}{Point to reflect}

\item{bounds}{Numeric vector of length 2 containing lower and upper bounds for where the root is}
}
\value{
Numeric value representing the reflected point
}
\description{
Implement the reflection of Coleman and Li (1994)
}
\details{
Implements the formula \code{min(wt,2*(ub - lb) - wt) + lb} where
\code{wt = abs(yt - lb) mod (2*(ub-lb))}
and \code{lb = min(bounds),ub = max(bounds)}.
}
