% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{summary.semibmd}
\alias{summary.semibmd}
\alias{print.summary.semibmd}
\title{Summary method for \code{semibmd} objects}
\usage{
\method{summary}{semibmd}(object, ...)

\method{print}{summary.semibmd}(x, digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{semibmd} returned by \code{semibmd::benchmark_dose}.}

\item{...}{Not used.}

\item{x}{Object of class \code{summary.semibmd} to print.}

\item{digits}{Number of digits to round bmd(l) for printing.}
}
\value{
Object of class \code{summary.semibmd}: a list with summary information, for printing
}
\description{
Summary method for \code{semibmd} objects returned by \code{semibmd::benchmark_dose}.
Combines the summary output of \code{scam/gam} with the benchmark dosing information
computed by \code{benchmark_dose}.
}
\details{
This function first calls the appropriate \code{summary} method for the
dose-response model, and then adds information about the benchmark dose and BMDL.
}
