% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{benchmark_dose}
\alias{benchmark_dose}
\title{Semi-parametric benchmark dosing}
\usage{
benchmark_dose(
  formula,
  data,
  exposure,
  x0 = 0,
  p0 = 0.05,
  BMR = 0.05,
  BMDL = c("all", "score", "delta"),
  monotone = TRUE,
  verbose = FALSE,
  eps = 1e-06,
  maxitr = 100,
  boot = 0
)
}
\arguments{
\item{formula}{formula passed to \code{scam::scam} or \code{mgcv::gam}}

\item{data}{\code{data.frame} passed to \code{scam::scam} or \code{mgcv::gam}}

\item{exposure}{Character vector naming the exposure variable in \code{formula}}

\item{x0}{Baseline exposure, default \code{0}}

\item{p0}{Baseline response}

\item{BMR}{Benchmark response}

\item{BMDL}{character: what kind of BMDL to return? \code{score} gives the score-based inversion lower bound;
\code{delta} gives the delta method-based lower bound. \code{all} gives all available options.}

\item{monotone}{Logical: should a monotone generalized additive model be used as the dose-response model?
Default \code{TRUE} uses \code{scam::scam}; setting to \code{FALSE} uses \code{mgcv::gam}.}

\item{verbose}{Logical, print progress and diagnostic information for debugging? Default \code{FALSE}.}

\item{eps}{Numerical tolerance}

\item{maxitr}{Maximum number of iterations. Default is fairly small, since when the problem is bounded it usually converges quickly.}

\item{boot}{Nonegative integer, how many bootstrap iterations to do for the parametric bootstrap BMDL? Default of \code{0}
means no bootstrapping.}
}
\description{
Benchmark dosing based on semi-parametric (monotone) generalized additive models
}
\note{
The BMDL assigned to the \code{bmdl} output object slot, and accessed via \code{get_bmd()} etc,
is always the score based one; the others exist only for comparison and diagnostics, and will be included
in the object's \code{info} slot.
}
