% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd-main.R
\name{bounded_newton}
\alias{bounded_newton}
\title{Reflected Newton's method}
\usage{
bounded_newton(g, bounds, xt = NULL, eps = 1e-06, maxitr = 10)
}
\arguments{
\item{g}{Function to find the root of. Passed to \code{rlang::as_function}.}

\item{bounds}{Numeric vector of length 2 containing lower and upper bounds for where the root is}

\item{xt}{Starting value, numeric, default \code{NULL} sets this to \code{mean(bounds)}.}

\item{eps}{Numerical tolerance}

\item{maxitr}{Maximum number of iterations. Default is fairly small, since when the problem is bounded it usually converges quickly.}
}
\value{
A numeric value representing the root.
}
\description{
R function implementing Newton's method for root finding with reflection
}
\details{
This function solves \code{g(x)=0} for a solution satisfying \code{x>=min(bounds)},
\code{x<=max(bounds)}, and \code{|g(x)|<eps}. It uses numerical derivatives.
}
\examples{
bounded_newton(function(x) x,c(-1,1))
bounded_newton(~.x,c(-1,1))
bounded_newton("force",c(-1,1))

}
